+++ NewsApp/Controllers/AuthorsController.cs
@@ -22,11 +22,22 @@ namespace NewsApp.Controllers
             return Ok(author);
         }
 
+        [HttpGet]
+        [Route("profile")]
+        public async Task<IActionResult> GetAuthorForCurrentUser()
+        {
+            var userIdString = _httpContextAccessor.HttpContext!.User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            var userId = int.Parse(userIdString);
+            var author = await _authorRepository.GetAuthorByUserId(userId);
+            return Ok(author);
+        }
+
         [HttpPost]
         public async Task<IActionResult> CreateAuthorFromCurentUser(Author author)
         {
             var userIdString = _httpContextAccessor.HttpContext!.User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
             var userId = int.Parse(userIdString);
+            // TODO: add check if user is already author
             author.UserId = userId;
             var authorId = await _authorRepository.CreateAuthor(author);
             return Ok(authorId);
@@ -34,7 +45,13 @@ namespace NewsApp.Controllers
 
         [HttpPut]
         public async Task<IActionResult> UpdateAuthor(Author author)
-        {     
+        {
+            var userIdString = _httpContextAccessor.HttpContext!.User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            var userId = int.Parse(userIdString);
+            if (!await _authorRepository.IsUserAuthor(userId, author.Id))
+            {
+                return Forbid();
+            }
             await _authorRepository.UpdateAuthor(author);
             return Ok();
         }
@@ -43,6 +60,12 @@ namespace NewsApp.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> DeleteAuthor([FromRoute] int id)
         {
+            var userIdString = _httpContextAccessor.HttpContext!.User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            var userId = int.Parse(userIdString);
+            if (!await _authorRepository.IsUserAuthor(userId, id))
+            {
+                return Forbid();
+            }
             await _authorRepository.DeleteAuthor(id);
             return Ok();
         }
